requirements.txt
Flask
gunicorn==21.2.0
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Login==0.6.1
Flask-WTF==0.15.1
python-dotenv
Werkzeug==2.0.3
Jinja2==3.0.3
itsdangerous==2.0.1
click==8.0.1
SQLAlchemy==1.4.25
psycopg2-binary==2.9.1
requests==2.26.0
pytest==6.2.4
pytest-flask==1.2.0
pytest-cov==2.12.1
pytest-mock==3.6.1
pytest-xdist==2.3.0
pytest-html==3.1.1
pytest-timeout==1.4.2
pytest-randomly==3.10.0
pytest-sugar==0.9.4
pytest-asyncio==0.15.1
pytest-django==4.5.2

